restructured server
