+++ server/README.md
@@ -14,7 +14,7 @@ This directory contains the backend code for the Quiz App, built using Python Fa
 
 ```bash
 git clone https://github.com/yourusername/quiz-app.git
-cd quiz-app/backend
+cd quiz-generator/server
 ```
 
 ### 2. Create virtual enviroment
@@ -54,7 +54,7 @@ PORT=8000
 Start the FastAPI server with the following command:
 
 ```bash
-pipenv run fastapi dev app/main.py
+pipenv run fastapi dev main.py
 ```
 
 The API will be accessible at `http://localhost:8000/api`.
