+++ server/api/v1/crud/generate_quiz.py
@@ -1,9 +1,9 @@
 from typing import Any
 from fastapi import HTTPException
 
-from libs.model import QuizQuestionsModel
+from ....schemas.model import QuizQuestionsModel
 
-from ..mocks.mock_quiz_data import (
+from ..db.mock_quiz_data import (
     quiz_data_multiple_choice,
     quiz_data_open_ended,
     quiz_data_true_false
