+++ server/api/v1/crud/generate_txt.py
