+++ server/api/v1/crud/get_user_quiz_history.py
@@ -1,4 +1,4 @@
-from ..mocks import mock_quiz_history
+from ..db import mock_quiz_history
 
 def get_user_quiz_history (user_id: str)-> list:
     return mock_quiz_history
