--- server/app/app_store/__init__.py
@@ -1,2 +1,0 @@
-from .mocks import *
-from .functions import *
