+++ server/schemas/query/generate_quiz_query.py
