+++ server/schemas/query/query_patterns.py
